backend/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23

# Database & Migrations
alembic==1.13.0

# Authentication & Security
bcrypt==4.1.2
pyjwt==2.8.0
pyotp==2.9.0
cryptography==41.0.7

# Task Scheduling
apscheduler==3.10.4

# API & Integrations
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
